requirements.txt
discord.py==2.1.*
matplotlib==3.8.*
pandas==2.2.*
requests==2.31.*
urllib3==2.2.*
APScheduler==3.10.*
SQLAlchemy==2.0.*